backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.20
pydantic==2.9.2
pydantic-settings==2.4.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy==2.0.35
better-exceptions==0.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.23.3
httpx==0.27.2